src/requirements.txt
openai
fastapi[standard]
python-dotenv
